Tests fixed
Due probably to different timing in the startup of the app, tests would succeed under windows but not under the linux and macos system I use; in one test there was no waiting time for the app to come up and the subsequent call would fail because the server wasn't there. Also, scheduled tasks weren't cleaned at shutdown. This does not impact the app, they were issues of the tests.