Issue #29: Allow different path for db companion YAML file
A further issue is that (also before this change) a ':' is a special character, but of course it's also a normal component of an absolute path in windows. Moreover, the way it is now, the string after a second '`:`' is discarded.

This creates (at least) two kinds of problem:

- `--db D:\myDb.db` is parsed as database `D` and yaml `\myDb.db`;
- `--mem-db myMem:D:\myMem.yaml` is parsed as yaml file `D`.

The solution is to detect if the path system uses the '`:`' as drive separator, and determine if the first colon is a drive separator or not. Also, after the first "real" colon, return all the string as the "second part".